+++ tasks/ec2_node.js
@@ -6,6 +6,7 @@ var commands = require('./lib/commands.js');
 module.exports = function (grunt) {
     var sshTask = require('./lib/sshTask.js')(grunt);
     var tasks = [
+        { name: 'ec2_pm2_update', command: 'sudo npm update -g pm2', description: 'Updates pm2 to latest' },
         { name: 'ec2_node_list', command: 'sudo pm2 list', description: 'Returns output for `pm2 list`' },
         { name: 'ec2_node_monit', command: 'sudo pm2 monit', description: 'Runs `pm2 monit`' },
         { name: 'ec2_node_reload', command: commands.pm2_reload, description: 'Reloads app using `pm2 reload all`' },